+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ENHANCEMENTS:
 * compute: add `variables` attribute in `yandex_compute_instance_group` resource and data source
 * managed-kubernetes: add `node_ipv4_cidr_mask_size` attribute to `yandex_kubernetes_cluster` resource and data source
+* serverless: add `environment` attribute in `yandex_function` resource and data source
 
 BUG FIXES:
 * mdb: fix modifying `yandex_mdb_mysql_cluster` attribute `backup_window_start`